requirements.txt
asgiref==3.5.2
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
Deprecated==1.2.13
Django==4.1.4
fonttools==4.38.0
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.6.2
numba==0.56.4
numpy==1.22.4
packaging==22.0
pandas==1.5.2
patsy==0.5.3
Pillow==9.3.0
pmdarima==2.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
statsmodels==0.13.5
threadpoolctl==3.1.0
tzdata==2022.7
urllib3==1.26.13
wrapt==1.14.1
django-storages==1.12.3
psycopg2-binary==2.9.5
boto3==1.21.14